requirements.txt
numpy==1.23.5
pandas==1.5.3
plotly==5.14.1
scipy==1.10.1
faiss-cpu
pickle5
